Revert "Ready for a release related to HyperDiff improvements"

+++ crates/hyper_ast/src/store/nodes/compo.rs
@@ -16,7 +16,7 @@ pub struct Size(pub u32);
 pub struct SizeNoSpaces(pub u32);
 pub struct Height(pub u32);
 pub struct BytesLen(pub u32);
-pub struct LineCount(pub u32);
+pub struct LineCount(pub u16);
 pub struct VizCsCount(pub u32);
 
 pub struct HStruct(pub u32);
